requirements.txt
Django==6.0.1
djangorestframework==3.14.0
django-filter==24.1
djangorestframework-simplejwt==5.3.2
asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.3
PyJWT==2.8.1
